refactoring with MVC
